requirements.txt
opencv-python
pyzbar
pillow
python-barcode
qrcode
pywin32